Move Drover Island demo into its own file
In order to show multiple demos in the app, we separate out the current
island demo into its own class, and use it from our main window.

This will enable us to add more demos while keeping their code separate.
