+++ winui-drover-island/winui-drover-island/DroverIslandDemo.h
@@ -14,28 +14,28 @@
 #pragma once
 
 #include <winrt/Microsoft.UI.Xaml.h>
-#include <winrt/winui_drover_island.h>
-
-#include "DroverIslandDemo.h"
 
 namespace winui_drover_island {
 
-class WinUIWindow
-{
+class DroverIslandDemo {
 public:
-	WinUIWindow();
-
-	void create();
-	void addContent();
-	void show();
-
-	const winrt::Microsoft::UI::Xaml::Window& window() const {
-		return mWindow;
-	};
+	DroverIslandDemo();
+	winrt::Microsoft::UI::Xaml::UIElement createContent();
 
 private:
-	winrt::Microsoft::UI::Xaml::Window mWindow{ nullptr };
-	DroverIslandDemo mDroverIslandDemo;
+	enum class Type { None, Ellipse, DroverSample, Last = DroverSample };
+
+	void renderCanvasControl(Type);
+	Type pickNextControl();
+
+	winrt::Microsoft::UI::Xaml::Controls::Grid mGrid{ nullptr };
+	winrt::Microsoft::UI::Xaml::Controls::Button::Click_revoker mClickRevoker;
+	winrt::Microsoft::UI::Xaml::Controls::CheckBox::Checked_revoker mCheckedRevoker;
+	winrt::Microsoft::UI::Xaml::Controls::CheckBox::Unchecked_revoker mUncheckedRevoker;
+	Type mControlType = Type::None;
+	winrt::Microsoft::UI::Xaml::Controls::Border mCanvasContainer{ nullptr };
+	winrt::Microsoft::UI::Xaml::Controls::TextBlock mDescription{ nullptr };
+	bool mUseVSIS = false;
 };
 
-}  // namespace winui_drover_island
+}
