+++ winui-drover-island/winui-drover-island/winui-drover-island.vcxproj
@@ -103,6 +103,7 @@
   <ItemGroup>
     <ClInclude Include="CanvasControl.h" />
     <ClInclude Include="DroverIsland.h" />
+    <ClInclude Include="DroverIslandDemo.h" />
     <ClInclude Include="EllipseShape.h" />
     <ClInclude Include="GfxD2DDeviceManager.h" />
     <ClInclude Include="GfxUtils.h" />
@@ -118,6 +119,7 @@
   <ItemGroup>
     <ClCompile Include="CanvasControl.cpp" />
     <ClCompile Include="DroverIsland.cpp" />
+    <ClCompile Include="DroverIslandDemo.cpp" />
     <ClCompile Include="EllipseShape.cpp" />
     <ClCompile Include="GfxD2DDeviceManager.cpp" />
     <ClCompile Include="GfxUtils.cpp" />
